refactor(ir): split scope_outline_utils.h into a header/impl pair
`scope_outline_utils.h` had grown to 2,254 header-only lines with no .cpp, so
every one of its 7 includers parsed and codegen'd the whole file. Four of those
includers referenced nothing from it but `outline_utils::VarDefUseCollector`,
which is not even defined here — it was a back-compat re-export of
`var_collectors::VarDefUseCollector` left over from an unfinished migration.
`expand_mixed_kernel_pass.cpp` wrote both spellings, in the same file.

Three changes, no behavior change:

1. Finish the var_collectors migration. Rewrite all 22
   `outline_utils::VarDefUseCollector` spellings to `var_collectors::` and
   delete the re-export. Also delete `using transform_utils::Substitute;`,
   which was unused by the header and by all 7 includers, and which sat at
   `namespace ir` scope so it leaked `ir::Substitute` into every TU that
   transitively included this header. Removing the alias also restores
   misc-include-cleaner, which had been accepting the fat header as a valid
   symbol provider on the strength of the re-export. 7 includers -> 4.

2. Extract the deferred-wait contract into `deferred_wait_contract.h`. The
   `ContainsDeferredWait` / `kMaxDeferredConditionsPerWaiter` /
   `DeferredWaitContractValidator` cluster is self-contained (its only outside
   dependency is `transform_utils::EvalConstTripCount`) and is precisely what
   `expand_mixed_kernel_pass.cpp` needs. The namespace is preserved, so no call
   site changes. 4 includers -> 3.

3. Give the header a .cpp. `ScopeOutliner` was 1,308 of the file's lines, and
   six helper classes (StoreTargetCollector, ParamReadCollector,
   PostStoreAliasCollector, SplitAivModeSummaryFinder, StoreEvalToAssignMutator,
   UpwardExposedUseCollector, ~500 lines) had no user outside it. The helpers
   move into an anonymous namespace in the new TU and ScopeOutliner's members
   are out-of-lined, keeping the class declaration, `VarCollector`, and the
   `ScopeKindAbsenceVerifier` template in the header. The `VisitScopeKind`
   member template moves too — all five instantiation sites are inside the
   class, hence inside the new TU.

Net: header 2,254 -> 355 lines, includers 7 -> 3, header-lines-times-TUs
~15,800 -> ~2,000.

Also updates the `utils.md` shared-utility table in both languages
(`StoreTargetCollector` is no longer public; new row for the extracted header)
and two comments in `verify_aiv_split.cpp` that named the .h as where
`split_aiv` is stamped.
