+++ docs/en/dev/passes/utils.md
@@ -119,7 +119,8 @@ auto used = memref_collectors::CollectUsedBasePtrs(func->body_);
 | ------ | --------- |
 | `transform_utils.h` | `Substitute`, `CollectDefVars`, `FindYieldStmt`, `FlattenToStmts`, `IsComputeTensorOp` |
 | `loop_state_repair.h` | `BuildDefMap`, loop rebuild helpers, `StripDeadIterArgs` |
-| `scope_outline_utils.h` | `VarCollector`, `StoreTargetCollector`, `ScopeOutliner`, `ScopeKindAbsenceVerifier` |
+| `scope_outline_utils.h` | `VarCollector`, `ScopeOutliner`, `ScopeKindAbsenceVerifier` |
+| `deferred_wait_contract.h` | `ContainsDeferredWait`, `DeferredWaitContractValidator`, `kMaxDeferredConditionsPerWaiter` |
 | `auto_name_utils.h` | SSA name generation, rename maps, name parsing |
 | `parent_stmt_analysis.h` | Parent-child statement mapping |
 | `dead_code_elimination.h` | Dead code removal within functions |
