+++ src/ir/transforms/utils/loop_state_repair.cpp
@@ -25,8 +25,8 @@
 #include "pypto/ir/stmt.h"
 #include "pypto/ir/transforms/utils/dead_code_elimination.h"
 #include "pypto/ir/transforms/utils/mutable_copy.h"
-#include "pypto/ir/transforms/utils/scope_outline_utils.h"
 #include "pypto/ir/transforms/utils/transform_utils.h"
+#include "pypto/ir/transforms/utils/var_collectors.h"
 
 namespace pypto {
 namespace ir {
@@ -97,7 +97,7 @@ namespace {
 void CollectBodyRefsSkippingYield(const std::vector<StmtPtr>& stmts, std::unordered_set<const Var*>& refs) {
   for (const auto& stmt : stmts) {
     if (std::dynamic_pointer_cast<const YieldStmt>(stmt)) continue;
-    outline_utils::VarDefUseCollector collector;
+    var_collectors::VarDefUseCollector collector;
     collector.VisitStmt(stmt);
     auto all_refs = collector.GetAllVarRefs();
     refs.insert(all_refs.begin(), all_refs.end());
@@ -127,7 +127,7 @@ StmtPtr FixDanglingYieldStmt(const StmtPtr& stmt, const std::vector<IterArgPtr>&
 
     std::vector<ExprPtr> new_values;
     for (size_t i = 0; i < yield_stmt->value_.size(); ++i) {
-      outline_utils::VarDefUseCollector collector;
+      var_collectors::VarDefUseCollector collector;
       collector.VisitExpr(yield_stmt->value_[i]);
       bool has_undefined = std::any_of(collector.var_uses.begin(), collector.var_uses.end(),
                                        [&](const Var* ref) { return !defined_vars.count(ref); });
@@ -163,7 +163,7 @@ void PullDefinitionChain(const Var* var_ptr, const std::unordered_map<const Var*
 
   auto assign = std::dynamic_pointer_cast<const AssignStmt>(it->second);
   if (assign) {
-    outline_utils::VarDefUseCollector collector;
+    var_collectors::VarDefUseCollector collector;
     collector.VisitExpr(assign->value_);
     for (const Var* dep : collector.var_uses) {
       PullDefinitionChain(dep, def_map, already_defined, pulled, out);
@@ -185,7 +185,7 @@ std::vector<StmtPtr> StripDeadIterArgs(const std::vector<StmtPtr>& stmts) {
     if (i + 1 < stmts.size()) {
       suffix_refs[i] = suffix_refs[i + 1];
     }
-    outline_utils::VarDefUseCollector collector;
+    var_collectors::VarDefUseCollector collector;
     collector.VisitStmt(stmts[i]);
     auto all_refs = collector.GetAllVarRefs();
     suffix_refs[i].insert(all_refs.begin(), all_refs.end());
@@ -301,7 +301,7 @@ std::vector<StmtPtr> FixupIterArgInitValues(const std::vector<StmtPtr>& stmts,
     if (iter_args_ptr && !iter_args_ptr->empty()) {
       std::vector<StmtPtr> missing_defs;
       for (const auto& iter_arg : *iter_args_ptr) {
-        outline_utils::VarDefUseCollector collector;
+        var_collectors::VarDefUseCollector collector;
         collector.VisitExpr(iter_arg->initValue_);
         for (const Var* ref : collector.var_uses) {
           if (!defined_so_far.count(ref) && !pulled.count(ref)) {
@@ -317,7 +317,7 @@ std::vector<StmtPtr> FixupIterArgInitValues(const std::vector<StmtPtr>& stmts,
       result.insert(result.end(), missing_defs.begin(), missing_defs.end());
     }
 
-    outline_utils::VarDefUseCollector stmt_defs;
+    var_collectors::VarDefUseCollector stmt_defs;
     stmt_defs.VisitStmt(stmt);
     defined_so_far.insert(stmt_defs.var_defs.begin(), stmt_defs.var_defs.end());
 
@@ -351,7 +351,7 @@ std::vector<StmtPtr> FixupDanglingYieldValues(const std::vector<StmtPtr>& stmts)
       const auto& iter_args = for_stmt ? for_stmt->iter_args_ : while_stmt->iter_args_;
       const auto& body = for_stmt ? for_stmt->body_ : while_stmt->body_;
 
-      outline_utils::VarDefUseCollector body_def_collector;
+      var_collectors::VarDefUseCollector body_def_collector;
       body_def_collector.VisitStmt(body);
       auto all_defined = defined_so_far;
       all_defined.insert(body_def_collector.var_defs.begin(), body_def_collector.var_defs.end());
@@ -374,7 +374,7 @@ std::vector<StmtPtr> FixupDanglingYieldValues(const std::vector<StmtPtr>& stmts)
       result.push_back(stmt);
     }
 
-    outline_utils::VarDefUseCollector stmt_defs;
+    var_collectors::VarDefUseCollector stmt_defs;
     stmt_defs.VisitStmt(stmt);
     defined_so_far.insert(stmt_defs.var_defs.begin(), stmt_defs.var_defs.end());
   }
@@ -392,7 +392,7 @@ std::shared_ptr<const YieldStmt> TrailingYield(const std::vector<StmtPtr>& stmts
 
 /// True when every Var referenced by `expr` is in `defined`.
 bool ExprRefsAllDefined(const ExprPtr& expr, const std::unordered_set<const Var*>& defined) {
-  outline_utils::VarDefUseCollector collector;
+  var_collectors::VarDefUseCollector collector;
   collector.VisitExpr(expr);
   for (const Var* ref : collector.var_uses) {
     if (!defined.count(ref)) return false;
@@ -542,7 +542,7 @@ std::vector<StmtPtr> StripDanglingIfReturnVars(const std::vector<StmtPtr>& stmts
     }
 
     result.push_back(new_stmt);
-    outline_utils::VarDefUseCollector c;
+    var_collectors::VarDefUseCollector c;
     c.VisitStmt(new_stmt);
     outer_defined.insert(c.var_defs.begin(), c.var_defs.end());
   }
